expressions.py

+++ Expressions/practice.c
@@ -5,6 +5,6 @@ int main(void){
     x = 2;
     y = 40;
     w = 4365;
-    printf("%f", (float)  x*y);
+    printf("%f", x*y);
 return 0;
 }
